refactor(engine): Remove expressions __init__ reexport
